feat: Option for Generate Self-Signed Certificate
This adds the cert_ext option to allow specifying the file extension for
just the self-signed certificate (not the key). It currently defaults
to "pem", but I now prefer to use "crt" convention you see just about
everywhere.

This change is backwards compatible to prevent breaking existing workflows
for such a small change.